requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
APScheduler==3.9.1
args==0.1.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
clint==0.5.1
gitlab3==0.5.8
dacite==1.6.0
dill==0.3.5.1
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
multidict==6.0.2
pyee==9.0.4
python-gitlab==3.10.0
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
slack-machine==0.30.0
slack-sdk==3.19.1
sniffio==1.3.0
typing-extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
yarl==1.8.1

